test_scripts/requirements.txt
requests==2.31.0
python-socketio[client]==5.10.0
pytest==7.4.3
pytest-asyncio==0.21.1 